Placed API method into its own class: _RpcMethodsTesting.
